♻️ refactor: const -> constants로 폴더 이름 변경
