+++ src/common/constants/resourceUrl/index.ts
