+++ src/components/pages/quiz/Comment/index.tsx
@@ -1,4 +1,4 @@
-import { ICON_URL } from '@/common/const/';
+import { ICON_URL } from '@/common/constants';
 import clsx from 'clsx';
 import Image from 'next/image';
